skills/aem/edge-delivery-services/aem-doc-converter/scripts/requirements.txt
markdown>=3.5
beautifulsoup4>=4.12
python-docx>=1.1
python-pptx>=0.6.23
pytest>=8.0
